# gen/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
  /vehicles/{id}:
    delete:
      operationId: deleteVehicle
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - vehicle-controller
      x-accepts: application/json
    get:
      operationId: getVehicleById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - vehicle-controller
      x-accepts: '*/*'
    put:
      operationId: updateVehicle
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - vehicle-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /users/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - user-controller
      x-accepts: application/json
    get:
      operationId: getUserById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
    put:
      operationId: updateUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - user-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /advertisements/{id}:
    delete:
      operationId: deleteAdvertisement
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - advertisement-controller
      x-accepts: application/json
    get:
      operationId: getAdvertisementById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Advertisement'
          description: OK
      tags:
      - advertisement-controller
      x-accepts: '*/*'
    put:
      operationId: updateAdvertisement
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Advertisement'
          description: OK
      tags:
      - advertisement-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /admin-page/vehicles/{id}:
    delete:
      operationId: deleteVehicle_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getVehicleById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateVehicle_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /admin-page/users/{id}:
    delete:
      operationId: deleteUser_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getUserById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateUser_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /admin-page/advertisements/{id}:
    delete:
      operationId: deleteAdvertisement_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getAdvertisementById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateAdvertisement_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /vehicles:
    get:
      operationId: getVehicles
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehicleDto'
                type: array
          description: OK
      tags:
      - vehicle-controller
      x-accepts: '*/*'
    post:
      operationId: createVehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - vehicle-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /vehicle-photos:
    get:
      operationId: getAllPhotos
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehiclePhotoDto'
                type: array
          description: OK
      tags:
      - vehicle-photo-controller
      x-accepts: '*/*'
    post:
      operationId: addPhotoToVehicle
      parameters:
      - explode: true
        in: query
        name: vehicleId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - explode: true
        in: query
        name: isMain
        required: true
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: username
        required: true
        schema:
          type: string
        style: form
      requestBody:
        $ref: '#/components/requestBodies/inline_object'
        content:
          application/json:
            schema:
              properties:
                photo:
                  format: binary
                  type: string
              required:
              - photo
              type: object
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehiclePhoto'
          description: OK
      tags:
      - vehicle-photo-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /users/{userId}/favorites/{advertisementId}:
    delete:
      operationId: removeFromFavorites
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
    post:
      operationId: addToFavorites
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
  /users/register:
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - user-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /user-page/favorites/{advertisementId}:
    delete:
      operationId: removeFromFavorites_1
      parameters:
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-page-controller
      x-accepts: '*/*'
    post:
      operationId: addToFavorites_1
      parameters:
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-page-controller
      x-accepts: '*/*'
...
  /advertisements:
    get:
      operationId: getAdvertisements
      parameters:
      - explode: true
        in: query
        name: page
        required: true
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 50
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageAdvertisementDto'
          description: OK
      tags:
      - advertisement-controller
      x-accepts: '*/*'
    post:
      operationId: createAdvertisement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Advertisement'
          description: OK
      tags:
      - advertisement-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /admin-page/vehicles:
    get:
      operationId: getAllVehicles
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehicleDto'
                type: array
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    post:
      operationId: createVehicle_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /admin-page/vehicle-photos/{vehicleId}:
    post:
      operationId: addPhotoToVehicle_1
      parameters:
      - explode: false
        in: path
        name: vehicleId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: true
        in: query
        name: isMain
        required: true
        schema:
          type: boolean
        style: form
      requestBody:
        $ref: '#/components/requestBodies/inline_object_2'
        content:
          application/json:
            schema:
              properties:
                photo:
                  format: binary
                  type: string
              required:
              - photo
              type: object
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehiclePhotoDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /admin-page/users/register:
    post:
      operationId: createUser_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /users/{userId}/favorites:
    get:
      operationId: getFavorites
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  format: int64
                  type: integer
                type: array
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
...
components:
  requestBodies:
...
    inline_object:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object'
...
  schemas:
...
    UserDto:
      example:
        password: password
        role: ADMIN
        id: 0
        email: email
        username: username
      properties:
        id:
          format: int64
          type: integer
        username:
          maxLength: 20
          minLength: 3
          type: string
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 6
          type: string
        role:
          enum:
          - ADMIN
          - USER
          - GUEST
          type: string
      required:
      - email
      - password
      - username
      type: object
    AdvertisementDto:
      example:
        id: 0
        title: title
        userId: 6
      properties:
        id:
          format: int64
          type: integer
        title:
          maxLength: 100
          minLength: 10
          type: string
        userId:
          format: int64
          type: integer
      required:
      - title
      type: object
    Advertisement:
      example:
        id: 0
        title: title
        user:
          favorites:
          - 1
          - 1
          advertisements:
          - null
          - null
          password: password
          role: ADMIN
          credentialsNonExpired: true
          accountNonExpired: true
          id: 6
          email: email
          enabled: true
          authorities:
          - authority: authority
          - authority: authority
          username: username
          accountNonLocked: true
      properties:
        id:
          format: int64
          type: integer
        title:
          type: string
        user:
          $ref: '#/components/schemas/User'
      type: object
...